App#13: Implement ComputeX, ComputeY, SecondaryAxisCoordinate WIP, think about changing float to decimal in MotionSetup and other relevant places

+++ src/Topics/Physics.ElectricParticle.Logic/Physics.ElectricParticle.Logic/IPhysicsService.cs
@@ -1,15 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using ExtendedNumerics;
 
 namespace Physics.ElectricParticle.Logic
 {
     interface IPhysicsService
     {
-        double ComputeX(double seconds);
+		BigDecimal ComputeX(BigDecimal seconds);
 
-        double ComputeY(double seconds);
+		BigDecimal ComputeY(BigDecimal seconds);
 
-        float MaxT { get; }
+        BigDecimal MaxT { get; }
     }
 }
